+++ src/eval/mod.rs
@@ -32,9 +32,12 @@ pub use self::redirect::{Redirect, RedirectAction, RedirectEval,
 pub use self::redirect_or_cmd_word::{EvalRedirectOrCmdWord, EvalRedirectOrCmdWordError,
                                     RedirectOrCmdWord, eval_redirects_or_cmd_words,
                                     eval_redirects_or_cmd_words_with_restorer};
-pub use self::redirect_or_var_assig::{EvalRedirectOrVarAssig, EvalRedirectOrVarAssigError,
-                                      RedirectOrVarAssig, eval_redirects_or_var_assignments,
+#[allow(deprecated)]
+pub use self::redirect_or_var_assig::{EvalRedirectOrVarAssig, eval_redirects_or_var_assignments,
                                       eval_redirects_or_var_assignments_with_restorer};
+pub use self::redirect_or_var_assig::{EvalRedirectOrVarAssig2, EvalRedirectOrVarAssigError,
+                                      RedirectOrVarAssig, eval_redirects_or_var_assignments2,
+                                      eval_redirects_or_var_assignments_with_restorers};
 
 /// A trait for evaluating parameters.
 pub trait ParamEval<E: ?Sized> {