+++ src/eval/redirect_or_var_assig.rs
@@ -1,6 +1,6 @@
 use {CANCELLED_TWICE, POLLED_TWICE};
-use env::{AsyncIoEnvironment, FileDescEnvironment, RedirectEnvRestorer, RedirectRestorer,
-          VariableEnvironment};
+use env::{AsyncIoEnvironment, ExportedVariableEnvironment, FileDescEnvironment, RedirectEnvRestorer,
+          RedirectRestorer, VarEnvRestorer, VariableEnvironment, VarRestorer, UnsetVariableEnvironment};
 use error::{IsFatalError, RedirectionError};
 use eval::{Assignment, RedirectEval, WordEval};
 use future::{Async, EnvFuture, Poll};
@@ -91,6 +91,8 @@ type RedirectOrVarAssigFuture<RF, V, WF> = RedirectOrVarAssig<RF, Option<V>, Ass
 /// returned on successful evaluation. These will **not** be applied to the
 /// environment at any point as that is left up to the caller.
 #[must_use = "futures do nothing unless polled"]
+#[deprecated(note = "does not properly handle references to earlier assignments, \
+use `EvalRedirectOrVarAssig2` instead")]
 pub struct EvalRedirectOrVarAssig<R, V, W, I, E: ?Sized, RR = RedirectRestorer<E>>
     where R: RedirectEval<E>,
           V: Hash + Eq,
@@ -102,6 +104,7 @@ pub struct EvalRedirectOrVarAssig<R, V, W, I, E: ?Sized, RR = RedirectRestorer<E
     rest: I,
 }
 
+#[allow(deprecated)]
 impl<R, V, W, I, E: ?Sized, RR> fmt::Debug for EvalRedirectOrVarAssig<R, V, W, I, E, RR>
     where I: fmt::Debug,
           R: RedirectEval<E>,
@@ -122,6 +125,51 @@ impl<R, V, W, I, E: ?Sized, RR> fmt::Debug for EvalRedirectOrVarAssig<R, V, W, I
     }
 }
 
+/// A future which will evaluate a series of redirections and variable assignments.
+///
+/// All redirections will be applied to the environment. On successful completion,
+/// a `RedirectRestorer` will be returned which allows the caller to reverse the
+/// changes from applying these redirections. On error, the redirections will
+/// be automatically restored.
+///
+/// In addition, all evaluated variable names and values to be assigned will be
+/// evaluated and added to the environment. On successful completeion, a
+/// `VarRestorer` will be returned which allows the caller to reverse the
+/// changes from applying those assignments. On error, the assignments will be
+/// automatically restored.
+#[must_use = "futures do nothing unless polled"]
+pub struct EvalRedirectOrVarAssig2<R, V, W, I, E: ?Sized, RR, VR>
+    where R: RedirectEval<E>,
+          V: Hash + Eq,
+          W: WordEval<E>,
+{
+    redirect_restorer: Option<RR>,
+    var_restorer: Option<VR>,
+    current: Option<RedirectOrVarAssigFuture<R::EvalFuture, V, W::EvalFuture>>,
+    rest: I,
+}
+
+impl<R, V, W, I, E: ?Sized, RR, VR> fmt::Debug for EvalRedirectOrVarAssig2<R, V, W, I, E, RR, VR>
+    where I: fmt::Debug,
+          R: RedirectEval<E>,
+          R::EvalFuture: fmt::Debug,
+          V: Hash + Eq + fmt::Debug,
+          W: WordEval<E>,
+          W::EvalFuture: fmt::Debug,
+          W::EvalResult: fmt::Debug,
+          RR: fmt::Debug,
+          VR: fmt::Debug,
+{
+    fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
+        fmt.debug_struct("EvalRedirectOrVarAssig2")
+            .field("redirect_restorer", &self.redirect_restorer)
+            .field("var_restorer", &self.var_restorer)
+            .field("current", &self.current)
+            .field("rest", &self.rest)
+            .finish()
+    }
+}
+
 /// Create a a future which will evaluate a series of redirections and variable assignments.
 ///
 /// All redirections will be applied to the environment. On successful completion,
@@ -130,7 +178,11 @@ impl<R, V, W, I, E: ?Sized, RR> fmt::Debug for EvalRedirectOrVarAssig<R, V, W, I
 /// be automatically restored.
 ///
 /// In addition, all evaluated variable names and values to be assigned will be
-/// returned on successful evaluation. These will **not** be applied to the
+/// returned on successful evaluation. These will **not** be applied
+/// environment at any point as that is left up to the caller.
+#[allow(deprecated)]
+#[deprecated(note = "does not properly handle references to earlier assignments, \
+use `eval_redirects_or_var_assignments2` instead")]
 pub fn eval_redirects_or_var_assignments<R, V, W, I, E: ?Sized>(vars: I, env: &E)
     -> EvalRedirectOrVarAssig<R, V, W, I::IntoIter, E, RedirectRestorer<E>>
     where I: IntoIterator<Item = RedirectOrVarAssig<R, V, W>>,
@@ -153,7 +205,11 @@ pub fn eval_redirects_or_var_assignments<R, V, W, I, E: ?Sized>(vars: I, env: &E
 /// be automatically restored.
 ///
 /// In addition, all evaluated variable names and values to be assigned will be
-/// returned on successful evaluation. These will **not** be applied to the
+/// returned on successful evaluation. These will **not** be applied
+/// environment at any point as that is left up to the caller.
+#[allow(deprecated)]
+#[deprecated(note = "does not properly handle references to earlier assignments, \
+use `eval_redirects_or_var_assignments_with_restorers` instead")]
 pub fn eval_redirects_or_var_assignments_with_restorer<R, V, W, I, E: ?Sized, RR>(
     mut restorer: RR,
     vars: I,
@@ -184,6 +240,77 @@ pub fn eval_redirects_or_var_assignments_with_restorer<R, V, W, I, E: ?Sized, RR
     }
 }
 
+/// Create a a future which will evaluate a series of redirections and variable assignments.
+///
+/// All redirections will be applied to the environment. On successful completion,
+/// a `RedirectRestorer` will be returned which allows the caller to reverse the
+/// changes from applying these redirections. On error, the redirections will
+/// be automatically restored.
+///
+/// In addition, all evaluated variable names and values to be assigned will be
+/// evaluated and added to the environment. On successful completeion, a
+/// `VarRestorer` will be returned which allows the caller to reverse the
+/// changes from applying those assignments. On error, the assignments will be
+/// automatically restored.
+pub fn eval_redirects_or_var_assignments2<R, V, W, I, E: ?Sized>(vars: I, env: &E)
+    -> EvalRedirectOrVarAssig2<R, V, W, I::IntoIter, E, RedirectRestorer<E>, VarRestorer<E>>
+    where I: IntoIterator<Item = RedirectOrVarAssig<R, V, W>>,
+          R: RedirectEval<E>,
+          V: Hash + Eq,
+          W: WordEval<E>,
+          E: ExportedVariableEnvironment + FileDescEnvironment + UnsetVariableEnvironment,
+          E::FileHandle: FileDescWrapper,
+          E::VarName: Borrow<String> + Clone,
+          E::Var: Borrow<String> + Clone,
+{
+    eval_redirects_or_var_assignments_with_restorers(RedirectRestorer::new(), VarRestorer::new(), vars, env)
+}
+
+/// Create a a future which will evaluate a series of redirections and variable assignments.
+///
+/// All redirections will be applied to the environment. On successful completion,
+/// a `RedirectEnvRestorer` will be returned which allows the caller to reverse the
+/// changes from applying these redirections. On error, the redirections will
+/// be automatically restored.
+///
+/// In addition, all evaluated variable names and values to be assigned will be
+/// evaluated and added to the environment. On successful completeion, a
+/// `VarEnvRestorer` will be returned which allows the caller to reverse the
+/// changes from applying those assignments. On error, the assignments will be
+/// automatically restored.
+pub fn eval_redirects_or_var_assignments_with_restorers<R, V, W, I, E: ?Sized, RR, VR>(
+    mut redirect_restorer: RR,
+    mut var_restorer: VR,
+    vars: I,
+    env: &E
+) -> EvalRedirectOrVarAssig2<R, V, W, I::IntoIter, E, RR, VR>
+    where I: IntoIterator<Item = RedirectOrVarAssig<R, V, W>>,
+          R: RedirectEval<E>,
+          V: Hash + Eq,
+          W: WordEval<E>,
+          E: ExportedVariableEnvironment + FileDescEnvironment + UnsetVariableEnvironment,
+          E::FileHandle: From<FileDesc> + Borrow<FileDesc>,
+          E::VarName: Borrow<String>,
+          E::Var: Borrow<String>,
+          RR: RedirectEnvRestorer<E>,
+          VR: VarEnvRestorer<E>,
+{
+    let mut vars = vars.into_iter();
+
+    let (lo, hi) = vars.size_hint();
+    let size_hint = hi.unwrap_or(lo);
+
+    redirect_restorer.reserve(size_hint);
+    var_restorer.reserve(size_hint);
+
+    EvalRedirectOrVarAssig2 {
+        redirect_restorer: Some(redirect_restorer),
+        var_restorer: Some(var_restorer),
+        current: vars.next().map(|n| spawn(n, env)),
+        rest: vars,
+    }
+}
+
 fn spawn<R, V, W, E: ?Sized>(var: RedirectOrVarAssig<R, V, W>, env: &E)
     -> RedirectOrVarAssigFuture<R::EvalFuture, V, W::EvalFuture>
     where R: RedirectEval<E>,
@@ -200,6 +327,93 @@ fn spawn<R, V, W, E: ?Sized>(var: RedirectOrVarAssig<R, V, W>, env: &E)
     }
 }
 
+// Allow resharing most of the state machine transitions between both implementations.
+// FIXME: This should go away once EvalRedirectOrVarAssig is fully deprecated.
+macro_rules! poll_impl {
+    ($self:ident, $env:expr, $var_insert:expr, $var_restore:expr, $get_vars:expr) => {{
+        let err;
+        loop {
+            if $self.current.is_none() {
+                $self.current = $self.rest.next().map(|next| spawn(next, $env));
+            }
+
+            let key_val = match $self.current {
+                Some(ref mut cur) => match *cur {
+                    RedirectOrVarAssig::Redirect(ref mut r) => {
+                        let action = match r.poll($env) {
+                            Ok(Async::Ready(action)) => action,
+                            Ok(Async::NotReady) => return Ok(Async::NotReady),
+                            Err(e) => {
+                                err = Some(EvalRedirectOrVarAssigError::Redirect(e.into()));
+                                break;
+                            },
+                        };
+
+                        let redirect_restorer = $self.redirect_restorer.as_mut()
+                            .take()
+                            .expect(POLLED_TWICE);
+                        match redirect_restorer.apply_action(action, $env) {
+                            Ok(()) => {},
+                            Err(e) => {
+                                err = Some(EvalRedirectOrVarAssigError::Redirect(
+                                    RedirectionError::Io(e, None).into()
+                                ));
+                                break;
+                            }
+                        }
+
+                        None
+                    },
+
+                    RedirectOrVarAssig::VarAssig(ref mut key, ref mut val) => {
+                        let val = match val.as_mut() {
+                            None => String::new().into(),
+                            Some(f) => match f.poll($env) {
+                                Ok(Async::Ready(val)) => val,
+                                Ok(Async::NotReady) => return Ok(Async::NotReady),
+                                Err(e) => {
+                                    err = Some(EvalRedirectOrVarAssigError::VarAssig(e));
+                                    break;
+                                },
+                            },
+                        };
+
+                        let key = key.take().expect(POLLED_TWICE).into();
+                        Some((key, val))
+                    },
+                },
+
+                None => {
+                    err = None;
+                    break
+                },
+            };
+
+            // Ensure we don't poll again
+            $self.current = None;
+
+            if let Some((key, val)) = key_val {
+                ($var_insert)($self, $env, key, val);
+            }
+        }
+
+        let mut restorer = $self.redirect_restorer.take().expect(POLLED_TWICE);
+
+        match err {
+            Some(e) => {
+                restorer.restore($env);
+                ($var_restore)($self, $env);
+                Err(e)
+            },
+            None => {
+                let vars = ($get_vars)($self);
+                Ok(Async::Ready((restorer, vars)))
+            },
+        }
+    }}
+}
+
+#[allow(deprecated)]
 impl<R, V, W, I, E: ?Sized, RR> EnvFuture<E> for EvalRedirectOrVarAssig<R, V, W, I, E, RR>
     where I: Iterator<Item = RedirectOrVarAssig<R, V, W>>,
           R: RedirectEval<E, Handle = E::FileHandle>,
@@ -215,77 +429,15 @@ impl<R, V, W, I, E: ?Sized, RR> EnvFuture<E> for EvalRedirectOrVarAssig<R, V, W,
     type Item = (RR, HashMap<V, W::EvalResult>);
     type Error = EvalRedirectOrVarAssigError<R::Error, W::Error>;
 
+    #[allow(redundant_closure_call)]
     fn poll(&mut self, env: &mut E) -> Poll<Self::Item, Self::Error> {
-        let err;
-        loop {
-            if self.current.is_none() {
-                self.current = self.rest.next().map(|next| spawn(next, env));
-            }
-
-            match self.current {
-                Some(ref mut cur) => match *cur {
-                    RedirectOrVarAssig::Redirect(ref mut r) => {
-                        let action = match r.poll(env) {
-                            Ok(Async::Ready(action)) => action,
-                            Ok(Async::NotReady) => return Ok(Async::NotReady),
-                            Err(e) => {
-                                err = Some(EvalRedirectOrVarAssigError::Redirect(e.into()));
-                                break;
-                            },
-                        };
-
-                        let restorer = self.redirect_restorer.as_mut().take().expect(POLLED_TWICE);
-                        match restorer.apply_action(action, env) {
-                            Ok(()) => {},
-                            Err(e) => {
-                                err = Some(EvalRedirectOrVarAssigError::Redirect(
-                                    RedirectionError::Io(e, None).into()
-                                ));
-                                break;
-                            }
-                        }
-                    },
-
-                    RedirectOrVarAssig::VarAssig(ref mut key, ref mut val) => {
-                        let val = match val.as_mut() {
-                            None => String::new().into(),
-                            Some(f) => match f.poll(env) {
-                                Ok(Async::Ready(val)) => val.into(),
-                                Ok(Async::NotReady) => return Ok(Async::NotReady),
-                                Err(e) => {
-                                    err = Some(EvalRedirectOrVarAssigError::VarAssig(e));
-                                    break;
-                                },
-                            },
-                        };
-
-                        let key = key.take().expect(POLLED_TWICE).into();
-                        self.vars.insert(key, val);
-                    },
-                },
-
-                None => {
-                    err = None;
-                    break
-                },
-            }
-
-            // Ensure we don't poll again
-            self.current = None;
-        }
-
-        let mut restorer = self.redirect_restorer.take().expect(POLLED_TWICE);
-
-        match err {
-            Some(e) => {
-                restorer.restore(env);
-                Err(e)
-            },
-            None => {
-                let vars = mem::replace(&mut self.vars, HashMap::new());
-                Ok(Async::Ready((restorer, vars)))
-            },
-        }
+        poll_impl!(
+            self,
+            env,
+            |slf: &mut Self, _: &mut E, key, val| slf.vars.insert(key, val),
+            |_, _: &mut E| {},
+            |slf: &mut Self| mem::replace(&mut slf.vars, HashMap::new())
+        )
     }
 
     fn cancel(&mut self, env: &mut E) {
@@ -297,5 +449,53 @@ impl<R, V, W, I, E: ?Sized, RR> EnvFuture<E> for EvalRedirectOrVarAssig<R, V, W,
         });
 
         self.redirect_restorer.take().expect(CANCELLED_TWICE).restore(env);
+    }
+}
+
+impl<R, V, W, I, E: ?Sized, RR, VR> EnvFuture<E> for EvalRedirectOrVarAssig2<R, V, W, I, E, RR, VR>
+    where I: Iterator<Item = RedirectOrVarAssig<R, V, W>>,
+          R: RedirectEval<E, Handle = E::FileHandle>,
+          R::Error: From<RedirectionError>,
+          V: Hash + Eq,
+          W: WordEval<E>,
+          E: AsyncIoEnvironment + ExportedVariableEnvironment + FileDescEnvironment + UnsetVariableEnvironment,
+          E::FileHandle: From<FileDesc> + Borrow<FileDesc>,
+          E::VarName: Borrow<String> + From<V>,
+          E::Var: Borrow<String> + From<W::EvalResult>,
+          RR: RedirectEnvRestorer<E>,
+          VR: VarEnvRestorer<E>,
+{
+    type Item = (RR, VR);
+    type Error = EvalRedirectOrVarAssigError<R::Error, W::Error>;
+
+    #[allow(redundant_closure_call)]
+    fn poll(&mut self, env: &mut E) -> Poll<Self::Item, Self::Error> {
+        poll_impl!(
+            self,
+            env,
+            |slf: &mut Self, env: &mut E, key, val: W::EvalResult| {
+                slf.var_restorer.as_mut()
+                    .expect(POLLED_TWICE)
+                    .set_exported_var(key, val.into(), true, env);
+            },
+            |slf: &mut Self, env: &mut E| {
+                slf.var_restorer.as_mut()
+                    .expect(POLLED_TWICE)
+                    .restore(env);
+            },
+            |slf: &mut Self| slf.var_restorer.take().expect(POLLED_TWICE)
+        )
+    }
+
+    fn cancel(&mut self, env: &mut E) {
+        self.current.as_mut().map(|cur| match *cur {
+            RedirectOrVarAssig::Redirect(ref mut f) => f.cancel(env),
+            RedirectOrVarAssig::VarAssig(_, ref mut f) => {
+                f.as_mut().map(|f| f.cancel(env));
+            },
+        });
+
+        self.redirect_restorer.take().expect(CANCELLED_TWICE).restore(env);
+        self.var_restorer.take().expect(CANCELLED_TWICE).restore(env);
     }
 }